temperature_graf/requirements.txt
FastAPI
pandas
beautifulsoup4
requests
lxml
html5lib
python-dotenv
uvicorn
Jinja2